refactor: transferring functions to specific files and organizing code
