+++ public/scripts/index.js
@@ -1,24 +1,31 @@
 import '../stylesheet.css';
+import badgesData from '../data/badges.json'
 
 import * as domUtils from './dom-utils.js';
+import { groupIntoRows } from './utils.js';
 import newStorage from './storage.js';
 
 import initVirtualWindow from './virtual-window.js';
 import initCopyButtons from './copy.js';
 import initColorScheme from './color-scheme.js';
-import initOrdering from './ordering.js';
+import initOrdering, { sortBadges } from './ordering.js';
 import initSearch from './search.js';
 
 console.log('Build #DEVELOPMENT_BUILD#');
 document.body.classList.remove('no-js');
+const loader = document.getElementById('loader');
+if (loader) {
+  loader.remove();
+}
 
 const storage = newStorage(localStorage);
-const loader = document.getElementById('loader');
-loader.remove();
-
-initVirtualWindow(window, document, navigator, storage);
 initColorScheme(document, storage);
 initCopyButtons(window, document, navigator, storage);
-// const orderingControls = initOrdering(document, storage);
+const { scroller, virtualWindowContainer } = initVirtualWindow(window, document, navigator, storage);
+const orderingControls = initOrdering(document, storage, (orderType) => {
+  const columnsCount = domUtils.getColumnsCount(virtualWindowContainer);
+  const sortedData = sortBadges(Object.values(badgesData), orderType);
+  scroller.setItems(groupIntoRows(sortedData, columnsCount));
+});
 initSearch(window.history, document, orderingControls, domUtils);
